Refactor imports to use tisslm module
This commit refactors the Python codebase to use the new `quanta_tissu.tisslm` module instead of the old `quanta_tissu.quanta_tissu` module.

The following files were updated to change the import paths:
- `quanta_tissu/tisslm/run_tiss.py`
- `quanta_tissu/tisslm/tools.py`
- `tests/features/steps/test_model_integration_steps.py`
- `tests/features/steps/test_predict_steps.py`
- `tests/features/steps/test_tokenizer_steps.py`
- `tests/test_bdd.py`
- `tests/test_execution_engine.py`
- `tests/test_knowledge_base.py`
- `tests/test_kv_cache.py`
- `tests/test_layers.py`
- `tests/test_model.py`

This change is part of a larger refactoring effort to clean up the codebase and improve the project structure. 11 out of 12 identified files have been updated.

The file `tests/test_tokenizer.py` still contains the old import paths and needs to be updated.
