+++ quanta_tissu/tisslm/run_tiss.py
@@ -9,8 +9,8 @@ import os
 sys.path.append(os.getcwd())
 
 from quanta_tissu.scripts.tisslang_parser import TissLangParser, TissLangParserError
-from quanta_tissu.quanta_tissu.execution_engine import ExecutionEngine, ToolRegistry
-from quanta_tissu.quanta_tissu.tools import run_command, write_file, read_file, assert_condition, TissCommandError
+from quanta_tissu.tisslm.execution_engine import ExecutionEngine, ToolRegistry
+from quanta_tissu.tisslm.tools import run_command, write_file, read_file, assert_condition, TissCommandError
 
 def main():
     """
